pyproject.toml
[project]
name = "langchain-agents"
version = "0.1.0"
description = "Multi porpose agents usign Langchain"
authors = [
    {name = "Jesus Diez",email = "jesussebastiandiezplasencia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain[openai] (>=0.3.25,<0.4.0)",
    "chromadb (>=1.0.9,<2.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langchain-chroma (>=0.2.4,<0.3.0)",
    "langgraph (>=0.4.5,<0.5.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "unstructured (>=0.17.2,<0.18.0)",
    "markdown (>=3.8,<4.0)"  
]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[tool.poetry]
packages = [{include = "langchain_agents", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
